+++ doc/release_notes.rst
@@ -10,6 +10,11 @@ Release Notes
 PyPSA-WAL Upcoming Release
 ==========================
 
+* Add uranium as a generator, to see nuclear generation in the model
+  (https://github.com/open-energy-transition/pypsa-wal/pull/36).
+* Update Walloon-specific potentials for biomass (including imported, transported, and local production), 
+  onshore wind, and solar using custom potentials provided in ``data/walloon/custom_potentials.csv`` 
+  (https://github.com/open-energy-transition/pypsa-wal/pull/29).
 * Overrode costs for specified fuels and technologies provided by ``data/walloon/custom_costs_rc.csv`` 
   (https://github.com/open-energy-transition/pypsa-wal/pull/31).
 * Added Walloon-specific nuclear decommissioning logic: custom staging data now